Use the new storage system completely for radius search, use it for confidence scores in all the other pages.
